requirements.txt
anyjson==0.3.3
appdirs==1.4.0
cffi==1.11.0
chaussette==1.3.0
circus==0.14.0
circus-web==1.0.0
cryptography==1.7.2
decorator==4.0.10
defusedxml==0.5.0
diogobaeder.webfaction==0.0.5
Django==1.10.5
django-allauth==0.30.0
django-bootstrap3==7.1.0
django-debug-toolbar==1.6
django-enumfields==0.8.2
enum34==1.1.6
Fabric3==1.13.1.post1
flake8==3.2.1
greenlet==0.4.12
idna==2.6
iowait==0.2
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
keyring==10.4.0
keyrings.alt==2.3
Mako==1.0.6
MarkupSafe==0.23
mccabe==0.5.3
meinheld==0.6.1
model-mommy==1.3.1
oauthlib==2.0.1
packaging==16.8
paramiko==2.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.2
pipdeptree==0.9.0
prompt-toolkit==1.0.9
psutil==5.1.3
psycopg2==2.6.2
ptyprocess==0.5.1
pyasn1==0.3.6
pycodestyle==2.2.0
pycparser==2.18
pycryptodome==3.4.5
pyflakes==1.3.0
Pygments==2.1.3
pylibmc==1.5.1
pyparsing==2.1.10
python-memcached==1.58
python3-openid==3.0.10
pyzmq==16.0.2
requests==2.13.0
requests-oauthlib==0.7.0
SecretStorage==2.3.1
simplegeneric==0.8.1
simplejson==3.10.0
six==1.11.0
sorl-thumbnail==12.4a1
sqlparse==0.2.2
tomako==0.1.0.post1
TornadIO2==0.0.3
tornado==4.4.2
traitlets==4.3.1
wcwidth==0.1.7
wfcli==0.4
